+++ Changelog.md
@@ -1,7 +1,14 @@
 # UNRELEASED
 
+### Vernacular
+- New attributes `#[proof="begin"]`, `#[proof(begin_if="att")]`
+  and `#[proof="end"]` to `Elpi Export` to
+  declare commands starting/ending proofs
+
 ### API
 - New `coq.redflags.proj` to control "unfolding" of primitive projections
+- New `main-interp-proof` and `main-interp-qed` for commands starting/ending
+  proofs
 
 # [2.5.2] 29/4/2025